requirements.txt
Django==5.0.6
djangorestframework==3.15
djangorestframework-jsonapi
django-filter
django-extensions
dj_database_url==2.1.0
ruff==0.4.5
mysqlclient

pytest
pytest-django
django-dynamic-fixture
